Add basic fairstep class, with (empty) shacl validation. Get rid of cwltool etc due to ancient rdflib requirements.
