+++ fairworkflows/__init__.py
@@ -1,4 +1,6 @@
 from ._version import __version__
-from .fairstep import fairstep
+from .fairstep import FairStep
 from .nanopub import Nanopub
 from .manualstep import manualstep
+
+from .cwlstep import cwlstep
