+++ fairworkflows/fairworkflow.py
@@ -1,0 +1,10 @@
+import rdflib
+
+class FairWorkflow:
+
+    def __init__():
+        steps = []
+
+    def add_step():
+        
+
