+++ requirements.txt
@@ -1,10 +1,9 @@
+pyshacl
 click
 Jinja2
-rdflib<4.3.0,>=4.2.2
+rdflib<6.0.0,>=5.0.0
 flaky
 pytest
-scriptcwl
-cwltool==2.0.20200312183052
 cwlgen
 pyyaml
 requests
